Update to PR comments
