--- TAS_T3DAdapter/CRUD/Create.cs
@@ -1,79 +1,0 @@
-﻿/*
- * This file is part of the Buildings and Habitats object Model (BHoM)
- * Copyright (c) 2015 - 2020, the respective contributors. All rights reserved.
- *
- * Each contributor holds copyright over their respective contributions.
- * The project versioning (Git) records all such contribution source information.
- *                                           
- *                                                                              
- * The BHoM is free software: you can redistribute it and/or modify         
- * it under the terms of the GNU Lesser General Public License as published by  
- * the Free Software Foundation, either version 3.0 of the License, or          
- * (at your option) any later version.                                          
- *                                                                              
- * The BHoM is distributed in the hope that it will be useful,              
- * but WITHOUT ANY WARRANTY; without even the implied warranty of               
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the                 
- * GNU Lesser General Public License for more details.                          
- *                                                                            
- * You should have received a copy of the GNU Lesser General Public License     
- * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
- */
-
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using BH.oM.Base;
-using BHE = BH.oM.Environment;
-using BHM = BH.oM.Environment.MaterialFragments;
-using BHG = BH.oM.Geometry;
-using System.Runtime.InteropServices;
-using BH.Engine.Environment;
-using System.Text;
-using System.Threading.Tasks;
-using BHA = BH.oM.Architecture;
-using BH.oM.Reflection.Attributes;
-using System.ComponentModel;
-using BHP = BH.oM.Environment.Fragments;
-using BH.Engine.TAS;
-using BH.oM.Adapter;
-
-namespace BH.Adapter.TAS
-{
-    public partial class TasT3DAdapter : BHoMAdapter
-    {
-        /***************************************************/
-        /**** Protected Methods                         ****/
-        /***************************************************/
-
-        protected override bool ICreate<T>(IEnumerable<T> objects, ActionConfig actionConfig = null)
-        {
-            bool success = true;
-
-            if (typeof(IBHoMObject).IsAssignableFrom(typeof(T)))
-            {
-                //success = CreateCollection(objects as dynamic);
-            }
-
-            return success;
-        }
-
-        /***************************************************/
-
-        public static void ClearCOMObject(object Object)
-        {
-            if (Object == null) return;
-            int intrefcount = 0;
-            do
-            {
-                intrefcount = Marshal.FinalReleaseComObject(Object);
-            } while (intrefcount > 0);
-            Object = null;
-        }
-    }
-
-    /***************************************************/
-
-}
-
-
