+++ TAS_T3DAdapter/TAS_T3DAdapter.csproj
@@ -115,7 +115,6 @@
   <ItemGroup>
     <Compile Include="AdapterActions\Pull.cs" />
     <Compile Include="AdapterActions\Push.cs" />
-    <Compile Include="CRUD\Create.cs" />
     <Compile Include="CRUD\Read.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="TAST3DAdapter.cs" />
